Rotate methode Point usw auf double umgestellt, und Praktikum 14 erledigt, mit Point.rotate Test

+++ GLP Projekte/Shapes/trunk/src/Shapes/Drawable.java
@@ -24,6 +24,6 @@ public abstract class Drawable
     * @param dy - distanz in y
     * 
     */
-    public abstract Drawable move (int dx, int dy);
+    public abstract Drawable move (double dx, double dy);
     public abstract Drawable rotate(Point center, double phi);
 }
